requirements.txt
Eel==0.17.0
numpy==2.2.0
openai==1.57.1
PyPDF2==3.0.1
python-dotenv==1.0.1
python_docx==1.1.2
scikit_learn==1.6.0
spacy==3.8.2
transformers==4.46.3
